+++ pkgs/test_core/lib/src/runner/load_suite.dart
@@ -5,6 +5,8 @@
 import 'dart:async';
 
 import 'package:stack_trace/stack_trace.dart';
+// ignore: deprecated_member_use
+import 'package:test_api/scaffolding.dart' show Timeout;
 import 'package:test_api/src/backend/group.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/invoker.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/metadata.dart'; // ignore: implementation_imports
@@ -12,8 +14,6 @@ import 'package:test_api/src/backend/runtime.dart'; // ignore: implementation_im
 import 'package:test_api/src/backend/suite.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/suite_platform.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/test.dart'; // ignore: implementation_imports
-// ignore: deprecated_member_use
-import 'package:test_api/scaffolding.dart' show Timeout;
 
 import '../util/async.dart';
 import '../util/io_stub.dart' if (dart.library.io) '../util/io.dart';
@@ -43,7 +43,7 @@ final _timeout = Duration(minutes: 12);
 ///
 /// A suite is constructed with logic necessary to produce a test suite. As with
 /// a normal test body, this logic isn't run until [LiveTest.run] is called. The
-/// suite itself is returned by [suite] once it's avaialble, but any errors or
+/// suite itself is returned by [suite] once it's available, but any errors or
 /// prints will be emitted through the running [LiveTest].
 class LoadSuite extends Suite implements RunnerSuite {
   @override
@@ -116,7 +116,7 @@ class LoadSuite extends Suite implements RunnerSuite {
       // If the test is forcibly closed, let it complete, since load tests don't
       // have timeouts.
       invoker.onClose.then((_) => invoker.removeOutstandingCallback());
-    }, completer.future, path: path);
+    }, completer.future, path: path, ignoreTimeouts: config.ignoreTimeouts);
   }
 
   /// A utility constructor for a load suite that just throws [exception].
@@ -143,23 +143,27 @@ class LoadSuite extends Suite implements RunnerSuite {
   }
 
   LoadSuite._(String name, this.config, SuitePlatform platform,
-      void Function() body, this._suiteAndZone, {String? path})
+      void Function() body, this._suiteAndZone,
+      {required bool ignoreTimeouts, String? path})
       : super(
             Group.root(
                 [LocalTest(name, Metadata(timeout: Timeout(_timeout)), body)]),
             platform,
-            path: path);
+            path: path,
+            ignoreTimeouts: ignoreTimeouts);
 
   /// A constructor used by [changeSuite].
   LoadSuite._changeSuite(LoadSuite old, this._suiteAndZone)
       : config = old.config,
-        super(old.group, old.platform, path: old.path);
+        super(old.group, old.platform,
+            path: old.path, ignoreTimeouts: old.ignoreTimeouts);
 
   /// A constructor used by [filter].
   LoadSuite._filtered(LoadSuite old, Group filtered)
       : config = old.config,
         _suiteAndZone = old._suiteAndZone,
-        super(old.group, old.platform, path: old.path);
+        super(old.group, old.platform,
+            path: old.path, ignoreTimeouts: old.ignoreTimeouts);
 
   /// Creates a new [LoadSuite] that's identical to this one, but that
   /// transforms [suite] once it's loaded.